1791: Fix aria-controls Axe error on View 12-week retest page
